adds warning and consistency
