+++ packages/app-mobile/src/screens/ResetWarningScreen.tsx
@@ -1,23 +1,19 @@
-import { View } from "react-native";
-
 import { UI_RPC_METHOD_USER_ACCOUNT_LOGOUT } from "@coral-xyz/common";
 import { useBackgroundClient, useUser } from "@coral-xyz/recoil";
 import { useNavigation } from "@react-navigation/native";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import { WarningIcon } from "~components/Icon";
 import {
   DangerButton,
-  Header,
-  Margin,
   Screen,
   SecondaryButton,
-  SubtextParagraph,
   TwoButtonFooter,
 } from "~components/index";
 
+import { WarningHeader } from "~src/features/warning";
 import { useSession } from "~src/lib/SessionProvider";
 
-export function LogoutWarningScreen({ navigation }): JSX.Element {
+export function LogoutWarningScreen(): JSX.Element {
   const background = useBackgroundClient();
   const user = useUser();
 
@@ -36,7 +32,7 @@ export function LogoutWarningScreen({ navigation }): JSX.Element {
   );
 }
 
-export function ResetWarningScreen({ navigation }): JSX.Element {
+export function ResetWarningScreen(): JSX.Element {
   const { reset } = useSession();
 
   return (
@@ -44,9 +40,7 @@ export function ResetWarningScreen({ navigation }): JSX.Element {
       buttonTitle="Reset"
       title="Reset Backpack"
       subtext="This will remove all the user accounts you have created or imported. Make sure you have your existing secret recovery phrase and private keys saved."
-      onNext={async () => {
-        await reset();
-      }}
+      onNext={reset}
     />
   );
 }
@@ -62,26 +56,15 @@ function Warning({
   subtext: string;
   onNext: () => void;
 }): JSX.Element {
+  const insets = useSafeAreaInsets();
   const navigation = useNavigation();
   const onPressCancel = () => {
     navigation.goBack();
   };
 
   return (
-    <Screen
-      style={{
-        justifyContent: "space-between",
-      }}
-    >
-      <View>
-        <Margin vertical={24}>
-          <View style={{ alignSelf: "center" }}>
-            <WarningIcon />
-          </View>
-        </Margin>
-        <Header text={title} />
-        <SubtextParagraph>{subtext}</SubtextParagraph>
-      </View>
+    <Screen jc="space-between" style={{ marginBottom: insets.bottom }}>
+      <WarningHeader title={title} subtitle={subtext} />
       <TwoButtonFooter
         leftButton={<SecondaryButton label="Cancel" onPress={onPressCancel} />}
         rightButton={<DangerButton label={buttonTitle} onPress={onNext} />}
